push scattering modes down into dust mix hierarchy
