+++ SKIRT/core/ElectronMix.hpp
@@ -43,11 +43,6 @@ public:
         is MaterialType::Electrons. */
     MaterialType materialType() const override;
 
-    /** This function returns the scattering mode supported by this material mix, which is
-        ScatteringMode::MaterialPhaseFunction or ScatteringMode::SphericalPolarization depending on
-        the value of the \em includePolarization flag. */
-    ScatteringMode scatteringMode() const override;
-
     /** This function returns the value of the \em includePolarization flag, indicating whether the
         material mix supports polarization during scattering events or not. */
     bool hasPolarizedScattering() const override;
@@ -102,45 +97,6 @@ public:
         scattering without or with support for polarization depending on the user-configured \em
         includePolarization property. */
     void performScattering(double lambda, const MediumState* state, PhotonPacket* pp) const override;
-
-    //======== Scattering with material phase function =======
-
-public:
-    /** This function returns the value of the scattering phase function
-        \f$\Phi_\lambda(\cos\theta)\f$ at wavelength \f$\lambda\f$ for the specified scattering
-        angle cosine \f$\cos\theta\f$, where the phase function is normalized as \f[\int_{-1}^1
-        \Phi_\lambda(\cos\theta) \,\mathrm{d}\cos\theta =2.\f] It invokes the corresponding
-        function of the DipolePhaseFunction class; see there for more information. */
-    double phaseFunctionValueForCosine(double lambda, double costheta) const override;
-
-    /** This function generates a random scattering angle cosine sampled from the phase function
-        \f$\Phi_\lambda(\cos\theta)\f$ at wavelength \f$\lambda\f$. It invokes the corresponding
-        function of the DipolePhaseFunction class; see there for more information. */
-    double generateCosineFromPhaseFunction(double lambda) const override;
-
-    //======== Polarization through scattering by spherical particles =======
-
-public:
-    /** This function returns the value of the scattering phase function
-        \f$\Phi_\lambda(\theta,\phi)\f$ at wavelength \f$\lambda\f$ for the specified scattering
-        angles \f$\theta\f$ and \f$\phi\f$, and for the specified incoming polarization state. The
-        phase function is normalized as \f[\int\Phi_\lambda(\theta,\phi) \,\mathrm{d}\Omega
-        =4\pi.\f] It invokes the corresponding function of the DipolePhaseFunction class; see there
-        for more information. */
-    double phaseFunctionValue(double lambda, double theta, double phi, const StokesVector* sv) const override;
-
-    /** This function generates random scattering angles \f$\theta\f$ and \f$\phi\f$ sampled from
-        the phase function \f$\Phi_\lambda(\theta,\phi)\f$ at wavelength \f$\lambda\f$, and for the
-        specified incoming polarization state. It invokes the corresponding function of the
-        DipolePhaseFunction class; see there for more information. The results are returned as a
-        pair of numbers in the order \f$\theta\f$ and \f$\phi\f$. */
-    std::pair<double, double> generateAnglesFromPhaseFunction(double lambda, const StokesVector* sv) const override;
-
-    /** This function applies the Mueller matrix transformation for the specified wavelength
-        \f$\lambda\f$ and scattering angle \f$\theta\f$ to the given polarization state (which
-        serves as both input and output for the function). It invokes the corresponding function of
-        the DipolePhaseFunction class; see there for more information. */
-    void applyMueller(double lambda, double theta, StokesVector* sv) const override;
 
     //======== Temperature and emission =======
