+++ SKIRT/core/MaterialMix.cpp
@@ -55,53 +55,9 @@ bool MaterialMix::hasStochasticDustEmission() const
 
 ////////////////////////////////////////////////////////////////////
 
-MaterialMix::ScatteringMode MaterialMix::scatteringMode() const
-{
-    return ScatteringMode::HenyeyGreenstein;
-}
-
-////////////////////////////////////////////////////////////////////
-
 double MaterialMix::asymmpar(double /*lambda*/) const
 {
     return 0.;
-}
-
-////////////////////////////////////////////////////////////////////
-
-double MaterialMix::phaseFunctionValueForCosine(double /*lambda*/, double /*costheta*/) const
-{
-    return 1.;
-}
-
-////////////////////////////////////////////////////////////////////
-
-double MaterialMix::generateCosineFromPhaseFunction(double /*lambda*/) const
-{
-    return 2. * random()->uniform() - 1.;
-}
-
-////////////////////////////////////////////////////////////////////
-
-double MaterialMix::phaseFunctionValue(double /*lambda*/, double /*theta*/, double /*phi*/,
-                                       const StokesVector* /*sv*/) const
-{
-    throw FATALERROR("This function implementation should never be called");
-}
-
-////////////////////////////////////////////////////////////////////
-
-std::pair<double, double> MaterialMix::generateAnglesFromPhaseFunction(double /*lambda*/,
-                                                                       const StokesVector* /*sv*/) const
-{
-    throw FATALERROR("This function implementation should never be called");
-}
-
-////////////////////////////////////////////////////////////////////
-
-void MaterialMix::applyMueller(double /*lambda*/, double /*theta*/, StokesVector* /*sv*/) const
-{
-    throw FATALERROR("This function implementation should never be called");
 }
 
 ////////////////////////////////////////////////////////////////////
